+++ src/pyedb/configuration/configuration.py
@@ -115,9 +115,9 @@ class Configuration:
             self.cfg_data.nets.apply()
 
         # Configure components
-        if self.cfg_data.components:
+        """if self.cfg_data.components:
             for comp in self.cfg_data.components:
-                comp.apply()
+                comp.apply()"""
 
         # Configure padstacks
         if self.cfg_data.padstacks:
